+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add missing SQLCipher proguard rules for R8
 - Show progress indicator when logging out
 - Bump to latest alpha/beta version of Compose libraries
+- Add month and year picker dialog for Compose forms
 
 ### Changes